+++ app.js
@@ -1,44 +1,45 @@
 'use strict';
 
 require('dotenv').config();
-const cors = require("cors")
+const cors = require('cors');
 const express = require('express');
 const app = express();
 app.disable('x-powered-by');
 
-const {errorResponse} = require('./src/utils/responseHandler');
-const {statusCodes} = require('./src/utils/statusCodes');
-const {router} = require('./src/routes/index');
+const { errorResponse } = require('./src/utils/responseHandler');
+const { statusCodes } = require('./src/utils/statusCode');
+const { router } = require('./src/routes/index');
 
-
-app.use(cors({
-    origin: 'trustedwebsite.com' //TODO:After the development is finished, add the origin.
-}));
+app.use(
+  cors({
+    origin: 'trustedwebsite.com', //TODO:After the development is finished, add the origin.
+  }),
+);
 
 app.use(express.json());
-app.use(express.urlencoded({extended: false}));
+app.use(express.urlencoded({ extended: false }));
 
 app.use('/api/v1', router);
 
 app.use((req, res, next) =>
-    errorResponse({
-        code: statusCodes.STATUS_CODE_DATA_NOT_FOUND,
-        req,
-        res,
-        message: 'Route not found',
-    }),
+  errorResponse({
+    code: statusCodes.STATUS_CODE_DATA_NOT_FOUND,
+    req,
+    res,
+    message: 'Route not found',
+  }),
 );
 
 app.use((error, req, res, next) =>
-    errorResponse({
-        code: statusCodes.STATUS_CODE_FAILURE,
-        req,
-        res,
-        error,
-        message: error.message,
-    }),
+  errorResponse({
+    code: statusCodes.STATUS_CODE_FAILURE,
+    req,
+    res,
+    error,
+    message: error.message,
+  }),
 );
 
 app.listen(process.env.PORT || 4000, () => {
-    console.log(`server started running on port ${process.env.PORT || 3000}`);
+  console.log(`server started running on port ${process.env.PORT || 3000}`);
 });
