+++ src/routes/distribution.route.js
@@ -1,0 +1,10 @@
+const {
+  addDistribution,
+} = require('../controllers/distributions/distribution.controller');
+const {
+  validateDistributionInput,
+} = require('../controllers/distributions/distribution.middleware');
+
+module.exports = function (app) {
+  app.post('/distributions', validateDistributionInput, addDistribution);
+};
