+++ src/routes/index.js
@@ -1,7 +1,7 @@
 const router = require('express').Router();
 
 require('./sample.route')(router);
-require('./distributions.route')(router);
+require('./distribution.route')(router);
 
 module.exports = {
   router,
