+++ src/utils/constant.js
@@ -1,0 +1,1 @@
+exports.dbTables = {};
