--- src/utils/constants.js
@@ -1,3 +1,0 @@
-exports.dbTables = {
-
-};
